requirements.txt
biopython==1.85
et_xmlfile==2.0.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
